image.look returns the colors used by an image
This is a better abstraction than the previous `colorAlloc`. This one has a better metaphor, and reads better in code, and has sensible semantics. The name is a pun on "_look_ at the image to see what color is uses" and "the image has a certain _look_", making it both a noun and verb.

Tests added for current expected functionality