Remove colorMap and just use rgbmap on the palette
Remove the colorMap component. It is confusing and no one I show raster.js to understands it. Instead, just require a palette, and have the palette contain an rgbmap (the old colorMap) and entries (the old palette; optional). This changes many APIs and obsoletes my weird features and edge-cases.

Some features are broken and tests are disabled, they still need to be fixed.