Improve Windows support and update the docs installation guide
On Windows, run a postinstall script that will copy the SDL2.dll dynamic library from the devel source to the build directory. This is required so that running the native addon will successfully find the dynamic library.

Update docs to reflect that SDL is now hosted on github, and to clarify the steps to install the native addon.