Pngjs and jpgjs for loading images on the web, fixes Firefox
Without this, image would be loaded on the web by creating an html <img> element, setting its src to the filename, drawing this element to a canvas, then reading the canvas's buffer. This is buggy in Firefox and sometimes blurs the image data. Use pure js implementations instead to make image loading work consistently across Chrome, Firefox, and node.js. Breaks one unit test for now. The behavior is correct, bug the test is hard to fix at the moment.